+++ src/Pages/DisplayReviews.jsx
@@ -1,6 +1,6 @@
 import { useLoaderData } from "react-router-dom";
 import { LOCAL } from "../App";
-import ReviewDisplay from "../components/ReviewList";
+import ReviewDisplay from "../components/admin/ReviewList";
 
 export default function DisplayReviewsPage() {
     const data = useLoaderData();
