+++ src/Pages/DisplayUsers.jsx
@@ -1,5 +1,5 @@
 import { useLoaderData } from "react-router-dom";
-import UserList from "../components/UserList";
+import UserList from "../components/admin/UserList.jsx";
 import { getAuthToken } from "../components/auth/auth";
 import { getBackendUrl } from "../util/localUrlGeneration.js";
